--- demo/src/main/controller/NoteController.java
@@ -1,22 +1,0 @@
-@CrossOrigin("*")
-@RestController
-@RequestMapping("/notes")
-
-public class NoteController {
-
-    private final NoteRepository noteRepository;
-
-    public NoteController(NoteRepository noteRepository) {
-        this.noteRepository = noteRepository;
-    }
-
-    @PostMapping("/add")
-    public Note addNote(@RequestBody Note note) {
-        return noteRepository.save(note);
-    }
-
-    @GetMapping("/all")
-    public List<Note> getAllNotes() {
-         return noteRepository.findAll();
-    }
-}
