+++ demo/src/main/java/com/example/demo/repository/NoteRepository.java
@@ -1,0 +1,6 @@
+package com.example.demo.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import com.example.demo.model.Note;
+
+public interface NoteRepository extends JpaRepository<Note,Long>{}
