--- demo/src/main/repository/NoteRepository.java
@@ -1,3 +1,0 @@
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface NoteRepository extends JpaRepository<Note,Long>{}
