+++ demo/src/main/resources/static/index.html
@@ -1,0 +1,59 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <title>Note Saver — Neo-Brutalist</title>
+  <script src="https://cdn.tailwindcss.com"></script>
+</head>
+<body class="bg-yellow-100 text-black min-h-screen flex flex-col items-center p-6">
+  <div class="w-full max-w-xl bg-white border-4 border-black shadow-[8px_8px_0px_#000] p-6 rounded-md">
+    <h1 class="text-3xl font-extrabold uppercase mb-4 border-b-4 border-black pb-2">📝 Note Saver</h1>
+
+    <textarea id="noteInput"
+              class="w-full p-4 border-4 border-black text-lg font-mono bg-pink-200 rounded-md resize-none shadow-[4px_4px_0px_#000]"
+              placeholder="Write your note here..."></textarea>
+
+    <button onclick="addNote()"
+            class="mt-4 w-full bg-blue-400 text-black font-bold text-lg border-4 border-black px-4 py-2 rounded-md shadow-[4px_4px_0px_#000] hover:bg-blue-300 active:translate-x-1 active:translate-y-1">
+      ➕ Add Note
+    </button>
+
+    <h2 class="text-2xl font-bold mt-8 mb-2 border-b-4 border-black pb-1">📋 All Notes</h2>
+    <ul id="noteList" class="space-y-2"></ul>
+  </div>
+
+  <script>
+    const API_URL = "http://localhost:8080/notes";
+
+    async function addNote() {
+      const content = document.getElementById("noteInput").value.trim();
+      if (!content) return alert("Note cannot be empty.");
+
+      await fetch(`${API_URL}/add`, {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ content })
+      });
+
+      document.getElementById("noteInput").value = "";
+      loadNotes();
+    }
+
+    async function loadNotes() {
+      const response = await fetch(`${API_URL}/all`);
+      const notes = await response.json();
+
+      const noteList = document.getElementById("noteList");
+      noteList.innerHTML = "";
+      notes.forEach(note => {
+        const li = document.createElement("li");
+        li.className = "bg-green-300 p-3 rounded-md border-4 border-black shadow-[4px_4px_0px_#000] font-mono";
+        li.textContent = note.content;
+        noteList.appendChild(li);
+      });
+    }
+
+    window.onload = loadNotes;
+  </script>
+</body>
+</html>
